fix(convert model to mesh): Updated code with removal of deprecated function converting model to mesh.
